pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "dataproc-ml-python"
version = "0.1.0"
description = "A python library to ease MLOps for Dataproc customers"
readme = "README.md"
requires-python = ">=3.11"
# Aligned with dependencies in 2.3 DPGCE ML image
dependencies = [
    "google-cloud-aiplatform~=1.88.0",
    "google-cloud-storage~=2.19.0",
    "pandas~=2.1.4",
    "pyarrow~=16.1.0",
    "pyspark~=3.5.3",
    "tenacity~=8.5.0",
    "tensorflow~=2.18.0",
    "torch~=2.6.0"
]

[project.optional-dependencies]
test = [
    "pytest",
    "torchvision~=0.21.0",
    "pillow~=11.3.0",
]
dev = [
    "pyink",
]

[tool.pyink]
line-length = 80 # Default is 88

[tool.pytest.ini_options]
addopts = "--ignore=load/"
